chore: separation and documentation
