+++ lib/do_nothing/formatter/markdown.ex
@@ -1,4 +1,7 @@
 defmodule DoNothing.Formatter.Markdown do
+  @moduledoc """
+  Default implementations for the `m:DoNothing.Formatter` callbacks.
+  """
   alias DoNothing.Formatter
   @behaviour Formatter
